move serialization headers to detail folder
