+++ libs/common/src/CMakeLists.txt
@@ -15,7 +15,6 @@ file(GLOB HEADER_LIST CONFIGURE_DEPENDS
         "${LaunchDarklyCommonSdk_SOURCE_DIR}/include/launchdarkly/logging/*.hpp"
         "${LaunchDarklyCommonSdk_SOURCE_DIR}/include/launchdarkly/data_sources/*.hpp"
         "${LaunchDarklyCommonSdk_SOURCE_DIR}/include/launchdarkly/data_sources/persistence/*.hpp"
-        "${LaunchDarklyCommonSdk_SOURCE_DIR}/include/launchdarkly/data_sources/serialization/*.hpp"
 )
 
 # Automatic library: static or dynamic based on user config.
