+++ libs/common/src/serialization/json_errors.cpp
@@ -1,4 +1,4 @@
-#include <launchdarkly/serialization/json_errors.hpp>
+#include <launchdarkly/detail/serialization/json_errors.hpp>
 
 namespace launchdarkly {
